Refactor Python position Example
Second change for the example how to combine `position`
and `positionOffset` to make it more general.

The functions provided now are corrected in namings and usable
for general `record components` even outside of the `particle`
scope.

I still did not add doc strings since I try not to blow up the
example since it is already pretty large. The `allocate` option
is quite of a fancy thing in case one wants to transparently
allocate an ndarray to access it via `[]` later on but could
be removed.